services/api-backend/requirements.txt
# Requirements pour LiveKit v1.x - Migration Juin 2025
fastapi==0.104.1
gunicorn==21.2.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
PyYAML==6.0.1
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
requests==2.31.0
numpy>=1.26.0
soundfile==0.12.1
aiofiles>=24.0.0
pydantic>=2.5.0
PyJWT==2.8.0

# LiveKit - Mise à jour pour utiliser les dernières versions
livekit
livekit-agents
# livekit-api est maintenant inclus dans livekit

# Dépendances audio
webrtcvad
scipy==1.11.4
waitress==3.0.0
torch==2.2.2
torchaudio==2.2.2
pydub==0.25.1

# Dépendances supplémentaires pour LiveKit v1.x
aiortc>=1.5.0
av>=10.0.0
pyee>=11.0.0
protobuf>=4.23.0